Added Working Carousel through js and css, and manipulated json file to work as jsonp file
***NOTE***
I had to change the names of the images to include underscores in with them because I was getting a weird bug where once I compiled and ran the code through to the browser, the image sources would change to the image names but with the underscores and the browser would say it didn't find the images. I spent the majority of my time working on this bug, trying to figure out what was causing it, seeing if it was one of the gulp libraries, but the only thing I could find out is that it was most likely occurring somewhere after sending the code through to the browser serve. After changing the image names to what the browser was telling me it saw, and changing the image sources in the css file, it finally worked. I tested it out with different names and it seemed to only do this with a string similar to the image names'. I'm guessing this was some type of formatting thing where the image names inside the css would change to a formatted version, and the formatted version would have the underscore for some reason. This is why I needed to change the images names and the .gitignore. If this is reason for point deduction, I apologize, I tried everything I could think of and didn't find anything online that could help me in this situation.